+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Added weather info
-- 
+- Added show battery settings
+  
 ### Changed
 - Updated from SDK5 to SDK6